修改code review時建議的小地方
