+++ Gemfile
@@ -40,7 +40,7 @@ group :development do
   gem 'web-console', '>= 4.1.0'
   # Display performance information such as SQL time and flame graphs for each request in your browser.
   # Can be configured to work on production as well see: https://github.com/MiniProfiler/rack-mini-profiler/blob/master/README.md
-  gem 'rack-mini-profiler', '~> 2.0'
+  # gem 'rack-mini-profiler', '~> 2.0'
   gem 'listen', '~> 3.3'
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
@@ -66,8 +66,6 @@ gem 'mini_magick'
 gem "fog-aws"
 # ENV variables
 gem 'figaro', '~> 1.2'
-# bootstrap for rails
-gem 'bootstrap-sass', '~> 3.4', '>= 3.4.1'
 # make form easy
 gem 'simple_form', '~> 5.1'
 # redirect_post to external website
