+++ app/controllers/application_controller.rb
@@ -1,6 +1,6 @@
 class ApplicationController < ActionController::Base
 
-
+  # 搬到users去
   def after_sign_in_path_for(resource)
     podcasts_url(subdomain: "host")
   end
