+++ db/migrate/20210529061102_remove_audio_from_episodes.rb
@@ -1,0 +1,5 @@
+class RemoveAudioFromEpisodes < ActiveRecord::Migration[6.1]
+  def change
+    remove_column :episodes, :audio, :string
+  end
+end
