+++ db/migrate/20210529061124_remove_avatar_from_users.rb
@@ -1,0 +1,5 @@
+class RemoveAvatarFromUsers < ActiveRecord::Migration[6.1]
+  def change
+    remove_column :users, :avatar, :string
+  end
+end
